bug(tests) fix profile and comments tests  - Tests for profile and comments no longer cause other tests to fail. [Finishes #169257416]
